Add option to force kill process after timeout
The previous implementation of "stop()" would always wait indefinitely
for the child process to end, which meant if the child process got
hung and/or did not respond to the SIGTERM signal then "stop()" would
never return. This adds the ability to kill the child process after a
timeout to give more control to the parent process in a backwards
compatible way.